Update InvitationListener.php
Reverting changes - We have been testing on our Laravel 5.3 product and this is correct. Checking the data sent through to each argument confirm that the $event and $data (renamed back to $invitation) were correct. There are no release notes indicating listeners have changed in Laravel 5.4 - if there is and this causes a problem then we can look at how to overcome it, but the project must work in 5.3.